third_party/blink/web_tests/platform/generic/fast/forms/textarea/textarea-minlength-expected.txt

Tests for HTMLTextAreaElement.minLength behaviors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textArea.minLength is -1
PASS textArea.minLength is -1
PASS textArea.minLength is -1
PASS textArea.minLength is 1
PASS textArea.minLength is -1
PASS textArea.minLength is -1
PASS textArea.minLength is 1
PASS textArea.minLength is 256
PASS textArea.getAttribute("minlength") is "6"
PASS textArea.minLength = -1 threw exception IndexSizeError: Failed to set the 'minLength' property on 'HTMLTextAreaElement': The value provided (-1) is not positive or 0..
PASS textArea.getAttribute("minlength") is "6"
PASS textArea.minLength = 11 threw exception IndexSizeError: Failed to set the 'minLength' property on 'HTMLTextAreaElement': The minLength provided (11) is greater than the maximum bound (10)..
PASS textArea.getAttribute("minlength") is "6"
PASS textArea.minLength = 10; textArea.getAttribute("minlength") is "10"
PASS textArea.minLength is 0
PASS textArea.getAttribute("minlength") is "0"
PASS successfullyParsed is true

TEST COMPLETE
